requirements.txt
pandas~=2.0.3
python-dotenv~=1.0.0
SQLAlchemy~=2.0.27
sentry-sdk~=2.17.0
minio~=7.2.15
pangres~=4.2.1
gspread~=6.1.4